Resolved version conflict of file core.navigation.Zoom.java, AERoport optimization apply
